feat: move app details to rsc
