+++ src/app/(dashboard)/app-store/[id]/components/AppDetailsTabs/index.ts
@@ -1,0 +1,1 @@
+export { AppDetailsTabs } from './AppDetailsTabs';
