+++ src/app/(dashboard)/app-store/[id]/components/InstallForm/index.ts
@@ -1,0 +1,1 @@
+export { InstallForm, type FormValues } from './InstallForm';
