+++ src/app/(dashboard)/app-store/[id]/components/InstallModal/index.ts
