+++ src/app/(dashboard)/app-store/[id]/components/StopModal/index.ts
@@ -1,0 +1,1 @@
+export { StopModal } from './StopModal';
