+++ src/app/(dashboard)/app-store/[id]/components/UninstallModal/index.ts
@@ -1,0 +1,1 @@
+export { UninstallModal } from './UninstallModal';
