+++ src/app/actions/app-actions/update-app-config-action.ts
@@ -1,0 +1,36 @@
+'use server';
+
+import { z } from 'zod';
+import { db } from '@/server/db';
+import { action } from '@/lib/safe-action';
+import { revalidatePath } from 'next/cache';
+import { AppServiceClass } from '@/server/services/apps/apps.service';
+import { handleActionError } from '../utils/handle-action-error';
+
+const formSchema = z.object({}).catchall(z.any());
+
+const input = z.object({
+  id: z.string(),
+  form: formSchema,
+  exposed: z.boolean().optional(),
+  domain: z.string().optional(),
+});
+
+/**
+ * Given an app id and form, updates the app config
+ */
+export const updateAppConfigAction = action(input, async ({ id, form, domain, exposed }) => {
+  try {
+    const appsService = new AppServiceClass(db);
+
+    await appsService.updateAppConfig(id, form, exposed, domain);
+
+    revalidatePath('/apps');
+    revalidatePath(`/app/${id}`);
+    revalidatePath(`/app-store/${id}`);
+
+    return { success: true };
+  } catch (e) {
+    return handleActionError(e);
+  }
+});
