--- src/client/modules/Apps/components/InstallForm/index.ts
@@ -1,1 +1,0 @@
-export { InstallForm } from './InstallForm';
