--- src/client/modules/Apps/pages/AppDetailsPage/index.ts
@@ -1,1 +1,0 @@
-export { AppDetailsPage } from './AppDetailsPage';
