+++ src/lib/helpers/castAppConfig.ts
