+++ src/server/services/apps/apps.service.test.ts
@@ -2,8 +2,8 @@ import fs from 'fs-extra';
 import waitForExpect from 'wait-for-expect';
 import { TestDatabase, clearDatabase, closeDatabase, createDatabase } from '@/server/tests/test-utils';
 import { faker } from '@faker-js/faker';
-import { castAppConfig } from '@/client/modules/Apps/helpers/castAppConfig';
 import { waitUntilFinishedMock } from '@/tests/server/jest.setup';
+import { castAppConfig } from '@/lib/helpers/castAppConfig';
 import { AppServiceClass } from './apps.service';
 import { EventDispatcher } from '../../core/EventDispatcher';
 import { getAllApps, getAppById, updateApp, createAppConfig, insertApp } from '../../tests/apps.factory';
